WC-987/fix: 파일명 typo 수정
