+++ wanchcoach/src/app/util/controller/medicalController.tsx
@@ -1,5 +1,6 @@
 import { Dayjs } from "dayjs";
 import fetchWithoutAuth from "../fetchWithoutAuth";
+import Cookies from "js-cookie";
 
 export interface HospitalDetailData {
   hospitalId: number;
@@ -40,6 +41,23 @@ interface queryData {
   location: LocationData;
 }
 
+interface MedicalKeywordSearchData {
+  keyword: string;
+  lng: string;
+  lat: string;
+}
+export interface MedicalKeywordResultData {
+  hospitalId: number;
+  name: string;
+  type: string;
+  address: string;
+}
+export interface PharmacyResultData {
+  pharmacyId: number;
+  name: string;
+  address: string;
+}
+
 export async function NearbyMedicalController(data: LocationData) {
   try {
     const url = `/medical/location-search?lng=${data.lng}&lat=${data.lat}`;
@@ -51,6 +69,21 @@ export async function NearbyMedicalController(data: LocationData) {
   } catch (error) {
     console.error("Error :", error);
     throw error; // 오류 처리
+  }
+}
+
+export async function MedicalKeywordSearchController(data: MedicalKeywordSearchData) {
+  try {
+    const url = `/medical/detail?keyword=${data.keyword}&lng=${data.lng}&lat=${data.lat}`;
+    const response = await fetchWithoutAuth(url, {
+      method: "GET",
+    });
+    console.log("keyword search success");
+    console.log(response);
+    return response;
+  } catch (error) {
+    console.error("error keyowrd search", error);
+    throw error;
   }
 }
